app/requirements.txt
Django==4.2.3
psycopg2-binary==2.9.6
django-ninja==0.19.1
django-cors-headers==3.13.0
pulp